Enable token persistance on wasm
Enable token persistance on wasm
