+++ src/Uno.AzureDevOps/Uno.AzureDevOps.Shared/Framework/Storage/SecureStorage.cs
@@ -1,11 +1,29 @@
 ﻿using System;
 using Newtonsoft.Json;
+using Windows.Security.Credentials;
+#if __WASM__
+using PasswordVault = Uno.AzureDevOps.Framework.Storage.PseudoPasswordVault;
+#endif
 
 namespace Uno.AzureDevOps.Framework.Storage
 {
+#if __WASM__
+	public sealed class PseudoPasswordVault : Windows.Security.Credentials.PasswordVault
+	{
+		public PseudoPasswordVault()
+			: base(new UnsecuredPersister(Uno.Foundation.WebAssemblyRuntime.InvokeJS("(function(){ return window.location.hostname; })();")?.ToLowerInvariant()))
+		{
+		}
+	}
+#endif
+
 	public partial class SecureStorage : ISecureStorage
 	{
 		private const string UadoResourceId = "Uno.AzureDevOps";
+
+#if !__ANDROID__
+        private readonly PasswordVault _passwordVault = new PasswordVault();
+#endif
 
 		public bool Delete(string key)
 		{
@@ -60,10 +78,58 @@ namespace Uno.AzureDevOps.Framework.Storage
 			}
 		}
 
-		partial void DeleteValue(string key, ref bool isDeleted);
+#if !__ANDROID__
+        private void DeleteValue(string key, ref bool isDeleted)
+		{
+			var credential = _passwordVault.Retrieve(UadoResourceId, key);
 
-		partial void GetValue(string key, ref string value);
+			if (credential != null)
+			{
+				_passwordVault.Remove(credential);
 
-		partial void SetValue(string key, string value, ref bool isSet);
+				isDeleted = true;
+			}
+			else
+			{
+				isDeleted = false;
+			}
+		}
+
+        private void GetValue(string key, ref string value)
+		{
+			try
+			{
+				var credential = _passwordVault.Retrieve(UadoResourceId, key);
+
+				credential.RetrievePassword();
+
+				value = credential?.Password;
+			}
+			catch
+			{
+				value = string.Empty;
+			}
+		}
+
+		private void SetValue(string key, string value, ref bool isSet)
+		{
+			try
+			{
+				_passwordVault.Add(new PasswordCredential()
+				{
+					Password = value,
+					Resource = UadoResourceId,
+					UserName = key
+				});
+
+				isSet = true;
+			}
+			catch (Exception e)
+			{
+				e.ToString();
+				isSet = false;
+			}
+		}
+#endif
 	}
 }
